backend/app/schemas.py
from pydantic import BaseModel, EmailStr
from datetime import datetime

class WaitlistCreate(BaseModel):
    email: EmailStr

class WaitlistResponse(BaseModel):
    id: int
    email: EmailStr
    created_at: datetime

    class Config:
        orm_mode = True
