backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
httpx
psycopg2-binary
requests
python-dotenv
